+++ Samples/Mapsui.Samples.CustomWidget/CustomWidget.cs
@@ -7,8 +7,8 @@ public class CustomWidget : IWidget
 {
     public HorizontalAlignment HorizontalAlignment { get; set; }
     public VerticalAlignment VerticalAlignment { get; set; }
-    public float MarginX { get; set; } = 20;
-    public float MarginY { get; set; } = 20;
+    public double MarginX { get; set; } = 20;
+    public double MarginY { get; set; } = 20;
     public MRect? Envelope { get; set; }
     public bool HandleWidgetTouched(Navigator navigator, MPoint position)
     {
@@ -17,7 +17,7 @@ public class CustomWidget : IWidget
     }
 
     public Color? Color { get; set; }
-    public int Width { get; set; }
-    public int Height { get; set; }
+    public double Width { get; set; }
+    public double Height { get; set; }
     public bool Enabled { get; set; } = true;
 }
